Make Generator a functional interface
Everything works yet, except... You can't use a generator multiple times, which
would be fine, except that Java caches "stateless" lambdas (even if the
functional interface is using a hack to give it state, but Java can't know
that).
